Adds allowed_postcodes model and specs
In order to fulfull the requirements of the service area model, the allowed_postcode model needed to be created to. Based on the requirements, I've assumed here that AllowedPostcodes generally belong to a service area for example Southwark. However in some cases they may not, for example the SH24 postcodes do not clearly belong to a service area. For this reason the service area relationship on allowed postcodes is optional and enforced at the database and application level.