Service area names are now unique
Adds specs and code to enforce case insensitive uniqueness constraint on ServiceArea.name property. Good news is that it broke some other tests and so had to edit validator specs who were trying to write a duplicate to the database. Uniqueness is enforced at both the application and database level.

- Closes #25