Adds postcode formatter as a concern to allowed post code
This works for the allowed post_code however it is not clear if it will be robust enough for EligibilityCheck. EligibilityCheck does not go through the same ActiveModel lifecycle as AllowedPostCode. When passing the value to the PostCodeValidator we want the value to be already be formatted. However, this formatter needs to run after validation to allow for constraints like presence and uniqueness to be checked first.